# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-commerce API",
        "description": "API REST completa do sistema de e-commerce. Todos os endpoints requerem autenticação via Bearer token (Sanctum), exceto os de autenticação pública.",
        "contact": {
            "email": "admin@ecommerce.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Listar produtos",
                "description": "List products with filters, search, sorting and pagination.",
                "operationId": "c821c7f9e41e08594ad4eae5a99e812e",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Busca por nome ou descrição",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filtrar por categoria",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "active",
                        "in": "query",
                        "description": "Filtrar por status ativo",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "in_stock",
                        "in": "query",
                        "description": "Filtrar com estoque disponível",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "name",
                                "price",
                                "created_at",
                                "quantity"
                            ]
                        }
                    },
                    {
                        "name": "sort_dir",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de produtos"
                    }
                }
            },
            "post": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Criar produto (admin)",
                "description": "Store a new product (admin only).",
                "operationId": "a59d7d76c37cb085c473698dba983938",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "quantity",
                                    "category_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Playstation 5"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 3550
                                    },
                                    "cost_price": {
                                        "type": "number",
                                        "example": 2800
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "min_quantity": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produto criado com sucesso"
                    },
                    "403": {
                        "description": "Acesso negado"
                    },
                    "422": {
                        "description": "Dados inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Exibir produto",
                "description": "Display a specific product.",
                "operationId": "900da7e5cb549dc28cfd19d90d1f6710",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do produto"
                    },
                    "404": {
                        "description": "Produto não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Atualizar produto (admin)",
                "description": "Update a product (admin only).",
                "operationId": "1ab61d88b4cfff271ad6afe7d8cf0388",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SuccessResponse"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produto atualizado com sucesso"
                    },
                    "403": {
                        "description": "Acesso negado"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Excluir produto (admin, soft delete)",
                "description": "Soft delete a product (admin only).",
                "operationId": "9dcbacba304004f42c516998d50fe6fa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto excluído com sucesso"
                    },
                    "403": {
                        "description": "Acesso negado"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/products/low-stock": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Produtos com estoque baixo (admin)",
                "description": "List products with low stock (admin only).",
                "operationId": "14b71c1518717dad5bca33e7a87a95f5",
                "responses": {
                    "200": {
                        "description": "Lista de produtos com estoque abaixo do mínimo"
                    },
                    "403": {
                        "description": "Acesso negado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "SuccessResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "data": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Mensagem de erro"
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}